refactor(homepage): inline organism/depth data into GAMES to eliminate circular imports
- Move organism/depth type definitions from organisms.ts to games.ts
- Inline organism and depth fields directly on each GAMES entry
- Remove module-load mutation of GAMES in organisms.ts
- Convert ORGANISM_BY_GAME to pure derivation from GAMES data
- Update test description to reflect new module structure
- Fix catalogNumber null safety with fallback to 0
- Improve link tap targets with min-height and padding
